get rid of utils.ts
